firefox: disable sandbox for 32-bit arm
When content sandbox is enabled, 32-bit arm build keeps crashing - at least the tabs only display this error. It looks like that due to some sandbox restrictions on this arch it is not able to load any fonts, that makes it give up on everything.

The build works also with sandbox, if MOZ_DISABLE_CONTENT_SANDBOX=1 environment variable is set.

For now, just disable sandbox for this arch, and investigate it later.

Also, remove "disable-elf-hack" option for x86, as it has no effect on its own.