firefox: remove meta-rust support
When the Rust version in oe-core is too old to compile Firefox, meta-lts-mixins
should be used. meta-rust uses a different target name scheme than the other layers,
so it adds some extra complexity to the recipes, which ultimately isn't required
now that the mixins layers have full support for Firefox.